requirements.txt
autocorrect==2.6.1
keyboard==0.13.5
pyperclip==1.8.2
python-dotenv==1.0.0
